Add files via upload
In this Project, where i have to keep track of how customers are liking the products of the company. You are keeping this track by capturing the number of likes against each Product by a customer. Write different types of CQL statements to insert and retrieve data from database.In this Project,where i have to keep a track of how customers are liking the products of the company. You are keeping this track by capturing the number of likes against each Product by a customer. Write different types of CQL statements to insert and retrieve data from database.
Skills: Cassandra · CQL statements · AstraDB